fix(sonar): cast para long na subtração do excesso de velocidade
Resolve java:S2184 no cálculo de percentual de excesso sem alterar o comportamento para os limites usados na prova.

+++ src/main/java/com/velsis/speed_violation_service/domain/service/ViolationEvaluationService.java
@@ -23,7 +23,7 @@ public class ViolationEvaluationService {
             return BigDecimal.ZERO.setScale(2, RoundingMode.UNNECESSARY);
         }
 
-        return BigDecimal.valueOf(consideredSpeed - speedLimit)
+        return BigDecimal.valueOf((long) consideredSpeed - speedLimit)
                 .multiply(BigDecimal.valueOf(100))
                 .divide(BigDecimal.valueOf(speedLimit), 2, RoundingMode.HALF_UP);
     }
